just utf-8 instead of reading it from options
'charset' - Displays the "Encoding for pages and feeds" set in Settings > Reading. This data is retrieved from the "blog_charset" record in the wp_options table. Note: In Version 3.5.0 and later, character encoding is no longer configurable from the Administration Panel. Therefore, this parameter always echoes "UTF-8", which is the default encoding of WordPress.

http://codex.wordpress.org/Function_Reference/bloginfo#Parameters


Signed-off-by: Norik Davtian <norik@bigemployee.com>